lemans: adsp: fix interrupt out port
The in port (3) is correct, but the out port is not. We couldn't find
the right info in ipcat, but it is available at the ADSP build in
core/systemdrivers/ipcc/image/adsp/hw/lemans/IPCCBSP.c, at the
"nInterrupt" field:

IPCCProtocolType IPCCProtocols[] =
{
  {
    .nProtocolId = IPCC_P_MPROC,
    .nNumSigs = 8,
    .nNumClients = NUM_CLIENTS,
    .nInterrupt = 153,
    .pClients = &IPCCClientMPROC[0],
    .bLegacySupported = 0,
    .bIPCCModeSupported = 1,
    .nPhysBaseAddress = 0x400000,
    .nOffset = 0x0,
    .bClearOnRecvEn = 1,
  }

With this, the calc test works on the ADSP. The ADSP still crashes
afterwards, so let's keep debugging.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>