tests: fix calc_test for local runs
The calc_tests use the QQVP_RESULTS_DIR env var, when set, to save
debugging ssh logs. However, if this var is unset, which is the case
when running locally, it tries to create the logs at the filesystem's
root. Let's fix that by only enabling log when the var is set.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>