tests: ensure sshd is up before running tests
Recently, we have been running into problems with CI where the
calc_tests timeout waiting for the ssh command to go through. The ssh
client seems to be able to connect, but not to deliver the command to be
executed. This _might_ be caused by the QNX sshd being not ready yet. So
let's try to establish a connection and run a simple command ("true")
with a small timeout before actually running the test, in order to
ensure sshd is up and fully prepared to handle connections.

Suggested-by: Brian Cain <bcain@quicinc.com>
Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>