Update the CCI broker stuff
- cci:get_cci_broker() looks up the 'current sc object' and then uses
that to get the 'current' broker.
The current sc object is the thing the sc kernel "ran" — which is likely
to be e.g. the CPU.
Where we use a cci broker, we dont end up with your broker, we end up
with the CPU's broker.
Basically before we had one broker so it was fine, but now we want to
have 'private' brokers in the models, to stop there being to many
parameters in one broker.
The way round this is to 'use' your local broker whenever you want to
'do' anything.
So we create a local broker inside each file which use it.