[QTOOL-100449]: add local option parsing and --version
Some QQVP options are only relevant to QQVP itself, not Qbox. This
includes the --help option that we had erroneously added to Qbox. Let's
move it to here, while also adding a --version option, as requested by
users. Removing --help from Qbox is left for a future MR (not that
QQVP's implementation will override/mask Qbox's, so this will work
nonetheless.)

Regarding --version, we use git-describe by default (which will use the
tags, if present, or the commit SHA), falling back to a local VERSION
file for tarball-originated builds. Failing to get both, compilation
will fail as well.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>