[QTOOL-94154]: CI: re-enable LRH test
With the longer login timeout from LRH, we can now re-enable this test
without the flakiness it once presented. We still leave the test out for
debug builds, though, as it takes too long to complete in those. Also,
TUN/TAP is not supported on OSX, and the only implementation I found
seems to be inactive since 2015 [1], so we disable the test in this case
as well. Running it on macOS gives: could not open /dev/tap0: No such
file or directory

[1]: https://tuntaposx.sourceforge.net/index.xhtml

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>