configs: red_hat/adb_debug: avoid red herring msg and use set -e
The line 'ip link show $bridge' sometimes show the message 'Device "br0"
does not exist'. This is actually expected, as we check if br0 already
exists and, if not, we create it. However, the message confuses us when
debugging red hat failures at CI, so let's redirect it to /dev/null.

While here, let's also use 'set -e' to make the script more robust and
facilitate debugging.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>