CMakeLists: fix installed binary issue on MacOS: 'Killed 9'
The installed vp on MacOS errors out immediately with 'Killed 9' due to
invalid codesign. This, in turn, seems to be caused by cmake's
installation methods and/or MacOS' cp behavior, which reuses the inode
and makes the codesing invalid. We add a workaround for that. Tested on
my personal mac.

Reported-by: Brian Cain <bcain@quicinc.com>
Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>