[QTOOL-104080] CI: fix wrong &&-chains causing undetected fails
GitLab runners execute before_script and script commands in a shell with
`set -e`, which exits on failures. However, there is a caveat with
&&-chains, as seen at [1] and [2]. In particular:

  "The shell does not exit if the command that fails is part of the
  command list immediately following a while or until keyword, part of
  the test in an if statement, part of any command executed in a && or
  || list except the command following the final && or ||, any command
  in a pipeline but the last, or if the command’s return status is being
  inverted with !."

This means that running something like "(set -e; false && true; echo
oops)", will actually end up printing "oops", as the shell will NOT exit
on the "false && true" statement (a bit confusing, IMO, but that is the
documented behavior). Many of our CI scripts actually rely on this WRONG
assumption, so we currently have jobs that should have stopped due to an
error proceeding normally.

Let's fix this using multi-line YAML statements ("-|" syntax), and let
GitLab's own `set -e` do what we were striving to do with the &&-chains.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/369229
[2]: https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>