Update conf.lua to support HVF
The conf.lua files describe the BSP contents to be loaded and for HVF
we want to omit the ARM-TF EL3 sw. So now the first boot stage in the
conf.lua files considers which mode we're using.

HVF changes for redhat config are not applied yet, we need to discuss
the appropriate bootstrapping for this mode.

Signed-off-by: Brian Cain <bcain@quicinc.com>
Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>


(cherry picked from commit c01631cf2451b8e5c960712f21642da9f4b4a668)