Fix -Wunused-variable warning
__gs_log_buffer is declared as 'static' which means that for each translation
unit it appears in, the compiler can exhaustively check for symbol references
to it.

This change to declare the array in the scope where it's used will eliminate
the warning.  Better still might be to create an object / archive that
allocates space with this name and refer to that one location.

    /local/mnt/workspace/vp/qbox/include/greensocs/gsutils/report.h: At global scope:
    /local/mnt/workspace/vp/qbox/include/greensocs/gsutils/report.h:17:13: warning: ‘gs::__gs_log_buffer’ defined but not used [-Wunused-variable]
       17 | static char __gs_log_buffer[100];
          |             ^~~~~~~~~~~~~~~

Signed-off-by: Brian Cain <bcain@quicinc.com>