mr: Make qemu_io_read/write public
This is required to read GICv3 dist and redist addresses from QQVP main.
The end goal is to create the memory region alias to be added as
subregion to the system memory. By reading at such address we can
trigger this indirectly, but I would like to discuss a better and more
direct way to achieve that.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>