interaction tests: fix issue to properly log ssh sessions
Interaction_test.py logs our ssh sessions to a directory that is
available at CI as an artifact. This should help us debug issues like
QTOOL-106439, where we got:

  subprocess.TimeoutExpired: Command 'ssh' timed out after 60 seconds

However, the log feature is currently broken due to a misuse of cmake's
if clause: 'if(<variable>)' will check whether the variable is defined
and evaluates to a true value. That's what we want, but the syntax is
'if(<variable>)', not 'if(${<variable>})'. Let's fix that. Also fix the
wrong variable name when constructing the log path.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>