[QTOOL-106439] tests: QCSubprocess: stop timer after process ends
We must not trigger TimeoutExpired exceptions in QCSubprocess if the
process has already finished, so let's check that. This seems to be part
of the reason behind the failures from QTOOL-106439: the ssh command
finishes successfully, but as we now continue testing with vp reboots
and new ssh cmds, the old ones end up timing out even though they've
succeeded (because QCSubprocess doesn't withdraw the timer).

While we are here, let's also change the timeout mechanism at
QCSubprocess to send SIGQUIT instead of SIGKILL, which gives qqvp and
other progs the chance to clean up.

Finally, it doesn't make much sense to open ssh connections with timeout
60s at _ensure_ssh_connects but then call success() with 20s as a
timeout. Instead, let's not restrict the QCSubprocess creation, and pass
the timeout parameter to success() itself.

Note: we still have ssh issues, but this "60 secs timeout" from
succeeded ssh commands should now be fixed.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>