luafile_tool: add --images-dir internal-only option
This option allows us to specify a dir to load images from, effectively
replacing QQVP_IMAGE_DIR. The dir will be inserted in a lua function for
conf.lua files to make use of implicitly (we don't want users to rely on
this new option, as it is meant only for development/testing).

We need top() for this added function, so it is added as well. This has
the extra benefit of allowing us to drop the many redundant top()
definitions in lua config files, and use this one that will now come by
default.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>