mem: Reduce system memory size by 1
Qbox uses a MMIO system_memory, effectively initializing it with
memory_region_init_io() instead of memory_region_init(). This is for
registering our callbacks with a MemoryRegionOps. Unfortunately we are
violating QEMU's expectations of a system_memory which should be a pure
"container" region.

Consequence of this is that by using UINT64_MAX as a size means "real"
size for a MMIO memory region rather than "all of the 64 bit address
space". Treating UINT64_MAX as real size ends up hitting an assert while
registering the memory region to KVM.

While redesigning Qbox memory management might be the right thing to do
to make QEMU happy, a quick and effective workaround is to just resize
the sistem memory size by 1.

More info can be found in the following discussion:
https://www.mail-archive.com/qemu-devel@nongnu.org/msg1001209.html

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>