cmake: Remove greensocs dependency
- The TARGET_LIBS should have the greensocs dependency with the call
of the macro gs_export() just before we add the platforms directory.
The problem was when we constructed the tarball, it asked the dependency
`Greensocs::qbox` which is created by gs_export(), but this macro
do nothing when we create the tarball.
So it will fix this kind of error: https://gitlab.qualcomm.com/qqvp/Qbox/-/jobs/2901633

Signed-off-by: tmarcero <tmarcero@quicinc.com>