Fix for quantum keeper locking
Due to a race condition in the quantum keeper (more specifically
`tlm_quantumkeeper_multithread`) synchronization sometimes ends up
deadlocking.

When `tlm_quantumkeeper_multithread::sync()` wakes up the SystemC thread
and starts waiting, what usually happens is that
`tlm_quantumkeeper_multithread::timehandler()` runs twice: Once for
unsuspending, so SystemC can update time, and once for suspending, after
SystemC time is up-to-date. The rough sequence of events, if SystemC
time is behind, is as follows:

1. `sync` calls `m_tick.notify` and `cond.wait` right after, to start
   waiting
2. `timehandler` runs the first time and unsuspends SystemC to update
   SystemC time and calls `cond.notify_all`
3. The predicate is evaluated the first time, because `cond.notify_all`
   was called and waiting continues, because time to sync is > 0
4. `timehandler` runs the second time and suspends SystemC because
   SystemC time has caught up. Then it calls `cond.notify_all` the
   second time
5. The predicate is evaluated the second time, but this time the
   predicate is true, because SystemC time caught up and execution
   continues.

Since `timehandler` is used as a callback and run in the SystemC thread,
there are occasions where this can cause a race condition. This happens
if the first evaluation of the predicate is interrupted and both
executions of `timehandler` finish before the second predicate
evaluation continues. The reason for this is that the predicate reads
SystemC time before it was updated and then misses `cond.notify_all`
from the second `timehandler`, after SystemC time is up-to-date. Hence
SystemC time isn't updated in the lambda.

Signed-off-by: Mark Burton <mburton@quicinc.com>
Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>