Fix CMakeLists.txt
Commit 46b494aa added changes to CMakeLists.txt related to the
recent libslirp issue. After that commit Gitlab CI started failing
because of what looks like a meson error on the surface:

`The command 'meson setup' failed with the output error: `

There is no output though.

The underlaying error is that meson is failing because the build
directory and the source directory don't exist. This can be an after
effect of `CPMAddPackage` downloading libslirp to `CPM_SOURCE_CACHE` if
it is set.

To fix this, we can use `<dependency>_SOURCE_DIR` and
`<dependency>_BINARY_DIR`. Those variables will be defined for a
dependency, after `CPMAddPackage` was called. In our case those are set
to: `libslirp_SOURCE_DIR` and `libslirp_BINARY_DIR`. This way source and
binary directories don't need to be hardcoded.

The reason for why there is no error message is that setting
`WORKING_DIRECTORY` masks the error for some reason. Maybe because it is
set to a directory that doesn't exist.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>