Combine lto blocks
Both blocks do the same thing. The compiler id check is unnecessary,
because `GS_ENABLE_LLD` is always set at the end of the block and the if
block right afterwards sets `-fuse-ld=lld`. That means if
`GS_ENABLE_LTO` is set, `-fuse-ld=lld` will also be set. But it's
possible to use lld and not lto.

Signed-off-by: Marco Liebel <mliebel@quicinc.com>