Update CCI repo version and remove RapidJSON
TLDR: Update CCI version to a version that uses a tested RapidJSON
version and get rid of manual download.

CMakeLists.txt wants to download the main branch of the SystemC CCI
repo, but the version is locked by package-lock.cmake to v1.0.1.
Removing the lock enables the use of more recent versions of SystemC
CCI, as is intended by CMakeLists.txt

One of the changes added after release v1.0.1 is the inclusion of
RapidJSON into the CCI codebase:
https://github.com/accellera-official/cci/pull/309
The rational behind it is, that RapidJSON doesn't release often and
adding it to the CCI repo makes sure that the commit used works with
CCI. Using the recommended version is a good thing imho. On top of that,
the manual download can be removed.

Signed-off-by: Marco Liebel <mliebel@quicinc.com>