SignalHandler: add dergister_cb function to deregister on_exit and handler callbacks
- when systemc sc_core::sc_stop() is called, the SignalHandler class destructor may not
be called before all other classes which uses the SignalHandler instance, this may cause
access to registered callbacks implemented as lambda functions capturing instances of the
destructed class. So, the deregister_cb for on_exit and signal handlers is used to remove
the registered callback at the destructor of the class with registered callback.
- use mutex to synchronize access to the various containers used in SignalHandler.
- add the deregister_cb to the destructors of all the classes using the SignalHandler
register_handler() and register_on_exit_cb().

Signed-off-by: Mahmoud Kamel <quic_mkamel@quicinc.com>