Switch SCP dependency to reference a SHA
The baseline referenced the 'main' branch.  But that way, our QBox builds
are not reproducible.  Any change to the SCP repo could cause existing
QBox tags to misbehave or fail to build correctly.

This sha represents the current 'main' as of the time this commit was
authored, so it shouldn't result in any change in behavior from the
baseline.

Now that we're switching to a specific sha, the gs_addpackage() could fail
if left with 'GIT_SHALLOW TRUE'.

See 969d395 (cmake: put GIT_SHALLOW to false, 2025-05-14) for another
example.

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>