python_binder: fix a problem with sc threads stack switching
- python 3.14 added a new stack overflow detection mechanism
which fails with systemc threads as they are implemented as user
space coroutines, the problem is very similar to what happens here
https://github.com/python/cpython/issues/139653
a new mechanism to run b_transport and bf_b_transport is introduced
to solve the problem.
- remove the dependency on numpy by making the tlm generic payload
object exposed to python implement the python buffer protocol. numpy
is still ok to use from python scripts, but it is not a requirement
for Qbox any more.
- Allow threads to run in python land by continously releasing the
python GIL from c++ side when it is not needed.
- Customize the python_binder test bench, py-uart.py and py-i2c.py
to accomodate the new changes.

Signed-off-by: Mahmoud Kamel <quic_mkamel@quicinc.com>